// apps/bot/src/types/baseEvent.ts
export type BaseEvent = {
	type: EventType
	timestamp: Date | string
	raw?: string
	chainID?: string
};

export enum EventType {
	PLAYER_WOUNDED = 'PLAYER_WOUNDED',
	PLAYER_DIED = 'PLAYER_DIED',
	PLAYER_CONNECTED = 'PLAYER_CONNECTED',
	PLAYER_DISCONNECTED = 'PLAYER_DISCONNECTED',
	ROUND_ENDED = 'ROUND_ENDED',
	NEW_GAME = 'NEW_GAME',
	SERVER_TICK_RATE = 'SERVER_TICK_RATE',
	PLAYER_POSSESS_CAMERA = 'PLAYER_POSSESS_CAMERA',
	PLAYER_UNPOSSESS_CAMERA = 'PLAYER_UNPOSSESS_CAMERA',
	ADMIN_BROADCAST = 'ADMIN_BROADCAST',
	FORCED_TEAM_CHANGE = 'FORCED_TEAM_CHANGE',
	CHAT_MESSAGE = 'CHAT_MESSAGE'
}